feat: filter jwt

+++ src/main/java/utils/contants/ENV.java
@@ -4,5 +4,5 @@ import config.EnvConfig;
 
 public interface ENV {
     String SECRET = EnvConfig.load().get("SECRET");
-    long ACCESS_TOKEN_EXPIRES_TIME = System.currentTimeMillis() + Long.parseLong(EnvConfig.load().get("ACCESS_TOKEN_DURATION"));
+    long ACCESS_TOKEN_EXPIRES_TIME = System.currentTimeMillis() + Integer.parseInt(EnvConfig.load().get("ACCESS_TOKEN_DURATION"));
 }
